chore: state 변경에 따른 구조 수정
project state에서 키 값 설정의 방식 변경과 여러 파일중 하나를 선택
하고 수정하는 경우를 고려하여 관련 컴포넌트들에서 state를 사용
하는 방식을 수정했습니다.
