+++ cocode/src/components/Project/BrowserV2/index.js
@@ -2,6 +2,7 @@ import React, { useEffect, useContext } from 'react';
 import * as Styled from './style';
 
 import ProjectContext from 'contexts/ProjectContext';
+// import { require } from 'bundler';
 
 function BrowserV2({ ...props }) {
 	const { project } = useContext(ProjectContext);
